pyproject.toml
[project]
name = "cn-pii-anonymization"
version = "0.1.0"
description = "中国个人信息脱敏库 - 识别和处理中国大陆个人身份信息"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "CN PII Team" }
]
keywords = ["pii", "anonymization", "privacy", "china", "chinese"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Text Processing",
]

dependencies = [
    "presidio-analyzer>=2.2",
    "presidio-anonymizer>=2.2",
    "spacy>=3.7",
    "fastapi>=0.109",
    "uvicorn>=0.27",
    "loguru>=0.7",
    "pillow>=10.0",
    "faker>=22.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "presidio-image-redactor>=0.0.50",
    "pytesseract>=0.3.10",
    "python-multipart>=0.0.22",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.2",
    "mypy>=1.8",
    "httpx>=0.26",
]
image = [
    "presidio-image-redactor>=0.0.50",
    "pytesseract>=0.3.10",
]

[project.scripts]
pii-anonymize = "cn_pii_anonymization.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cn_pii_anonymization"]

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "*.egg-info",
    ".mypy_cache",
    ".pytest_cache",
]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "N",
    "W",
    "UP",
    "B",
    "C4",
    "SIM",
    "TCH",
    "RUF",
]
ignore = [
    "E501",
    "B008",
    "RUF001",
    "RUF002",
    "RUF003",
]

[tool.ruff.lint.isort]
known-first-party = ["cn_pii_anonymization"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/cn_pii_anonymization"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]

[dependency-groups]
dev = [
    "pytest-benchmark>=5.2.3",
]
